refactor: split terminal auth helpers
