+++ src/backend/ssh/terminal.ts
@@ -3,14 +3,8 @@ import ssh2Pkg, {
   type Client as SSHClientType,
   type ClientChannel,
   type PseudoTtyOptions,
-  type ParsedKey,
-  type SignCallback,
-  type SigningRequestOptions,
-  type IdentityCallback,
 } from "ssh2";
-const { Client, BaseAgent, utils: ssh2Utils } = ssh2Pkg;
-import net from "net";
-import dgram from "dgram";
+const { Client, utils: ssh2Utils } = ssh2Pkg;
 import { SSH_ALGORITHMS } from "../utils/ssh-algorithms.js";
 import axios from "axios";
 import { getDb } from "../database/db/index.js";
@@ -33,76 +27,7 @@ import {
   attachOrCreateTmuxSession,
   waitForTmuxSession,
 } from "./tmux-helper.js";
-
-class MemoryAgent extends BaseAgent {
-  private key: ParsedKey;
-
-  constructor(key: ParsedKey) {
-    super();
-    this.key = key;
-  }
-
-  getIdentities(cb: IdentityCallback<ParsedKey>): void {
-    cb(null, [this.key]);
-  }
-
-  sign(
-    pubKey: ParsedKey | Buffer | string,
-    data: Buffer,
-    optionsOrCb: SigningRequestOptions | SignCallback,
-    cb?: SignCallback,
-  ): void {
-    const callback = typeof optionsOrCb === "function" ? optionsOrCb : cb!;
-    const options = typeof optionsOrCb === "function" ? {} : optionsOrCb;
-    try {
-      const algo =
-        options.hash === "sha256"
-          ? "rsa-sha2-256"
-          : options.hash === "sha512"
-            ? "rsa-sha2-512"
-            : undefined;
-      const signature = this.key.sign(data, algo);
-      callback(null, signature);
-    } catch (err) {
-      callback(err instanceof Error ? err : new Error(String(err)));
-    }
-  }
-}
-
-async function performPortKnocking(
-  host: string,
-  sequence: Array<{ port: number; protocol?: string; delay?: number }>,
-): Promise<void> {
-  for (const knock of sequence) {
-    const protocol = knock.protocol || "tcp";
-    const delay = knock.delay ?? 100;
-
-    await new Promise<void>((resolve) => {
-      if (protocol === "udp") {
-        const client = dgram.createSocket("udp4");
-        client.send(Buffer.alloc(0), knock.port, host, () => {
-          client.close();
-          resolve();
-        });
-      } else {
-        const socket = new net.Socket();
-        socket.once("connect", () => {
-          socket.destroy();
-          resolve();
-        });
-        socket.once("error", () => {
-          socket.destroy();
-          resolve();
-        });
-        socket.connect(knock.port, host);
-      }
-    });
-
-    if (delay > 0) {
-      await new Promise<void>((r) => setTimeout(r, delay));
-    }
-  }
-}
+import { MemoryAgent, performPortKnocking } from "./terminal-auth-helpers.js";
 
 interface ConnectToHostData {
   cols: number;
